separated the network from the trainer. added ability to initialize from a previous model with a bigger or smaller base size. need to add ability to initialize model with a different puzzle length and word length
