+++ baseline/experiments/base_size_experiment.py
@@ -4,7 +4,7 @@
 import csv
 
 # Ipmort the model and the generator
-from model.nn import Trainer
+from model.trainer import Trainer
 from model.generator import Generator
 
 
@@ -25,7 +25,7 @@ def train_model(base, training_size, length=2, choice=5, epochs=2000, batch_size
     trainer = Trainer(train_data, test_data, epochs, dimension)
 
     # run model on generated data
-    model = trainer.batch_train(1000)
+    model = trainer.batch_train()
 
     train_acc = trainer.evaluate(model, train_data[:200])
     test_acc = trainer.evaluate(model, test_data)
@@ -33,6 +33,7 @@ def train_model(base, training_size, length=2, choice=5, epochs=2000, batch_size
     return (base, train_acc, test_acc)
 
 def run_experiment():
+    print('experimenting with base size')
     for training_size in TRAINING_SIZES:
         results = []
         for base_size in BASE_SIZES:
@@ -41,6 +42,6 @@ def run_experiment():
         with open("baseline/results/base_size_experiment/base_size_experiment_training_size_" + str(training_size) + ".csv", "w") as f:
              writer = csv.writer(f)
              writer.writerows(results)
+
 if __name__ == '__main__':
-    print('experimenting with base size')
     run_experiment()
