+++ baseline/model/nn.py
@@ -7,14 +7,12 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import torch.optim as optim
-from model.generator import Generator
-from model.vectorize import buildVocab, makePuzzleVector, makePuzzleTarget, makePuzzleTargets, makePuzzleMatrix
 
 if torch.cuda.is_available():
-    print("trainer using gpu")
+    print("network using gpu")
     cuda = torch.device('cuda:0')
-    FloatTensor = torch.cuda.FloatTensor
-    LongTensor = torch.cuda.LongTensor
+    FloatTensor = torch.FloatTensor
+    LongTensor = torch.LongTensor
     def cudaify(model):
         model.cuda()
 else:
@@ -85,116 +83,5 @@ class MultiLayerClassifier(nn.Module):
         result.input_layer.weight = torch.nn.Parameter(input_layer_weights)
         return result
 
-
-
-class Trainer:
-
-    def __init__(self, train_data, test_data, epochs = 20, dimension = 300, num_hidden = 2, batch_size = 1000):
-        self.num_training_epochs = epochs
-        self.hidden_layer_size = dimension
-        self.num_hidden = num_hidden
-        self.num_choices = len(train_data[0][0])
-        self.train_data = train_data
-        self.test_data = test_data
-        self.vocab = buildVocab(self.train_data + self.test_data)
-        self.batch_size = batch_size
-
-    def train(self):
-        model = MultiLayerClassifier(self.vocab, self.num_choices, self.hidden_layer_size, self.num_hidden)
-        print(model)
-        loss_function = nn.NLLLoss()
-        optimizer = optim.Adam(model.parameters())
-        for epoch in range(self.num_training_epochs):
-            print('epoch {}'.format(epoch))
-            for instance, label in self.train_data:
-                model.zero_grad()
-                input_vec = makePuzzleVector((instance, label), self.vocab)
-                target = makePuzzleTarget(label)
-                log_probs = model(input_vec)
-                loss = loss_function(log_probs, target)
-                loss.backward()
-                optimizer.step()
-            train_acc = self.evaluate(model, self.train_data[:200])
-            test_acc = self.evaluate(model, self.test_data)
-            print('train: {:.2f}; test: {:.2f}'.format(train_acc, test_acc))
-        return model
-
-    def batch_train(self, model):
-        print(model)
-        loss_function = nn.NLLLoss()
-        batch_size = self.batch_size
-        optimizer = optim.Adam(model.parameters())
-        for epoch in range(self.num_training_epochs):
-            model.zero_grad()
-            batch = random.sample(self.train_data, batch_size)
-            input_matrix = makePuzzleMatrix(batch, model.vocab)
-            target = makePuzzleTargets([label for (_, label) in batch])
-            log_probs = model(input_matrix)
-            loss = loss_function(log_probs, target)
-            loss.backward()
-            optimizer.step()
-            if epoch % 100 == 0:
-                print('epoch {}'.format(epoch))
-                train_acc = self.evaluate(model, self.train_data[:200])
-                test_acc = self.evaluate(model, self.test_data)
-                print('train: {:.2f}; test: {:.2f}'.format(train_acc, test_acc))
-        return model
-
-
-    def evaluate(self, model, test_d):
-        """Evaluates the trained network on test data."""
-        word_to_ix = self.vocab
-        with torch.no_grad():
-            correct = 0
-            for instance, label in test_d:
-                input_vec = makePuzzleVector((instance, label), word_to_ix)
-                log_probs = model(input_vec)
-                probs = [math.exp(log_prob) for log_prob in log_probs.tolist()[0]]
-                ranked_probs = list(zip(probs, range(len(probs))))
-                response = max(ranked_probs)[1]
-                if response == label:
-                    correct += 1
-        return correct/len(test_d)
-
-#ALPHABET = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz'
-
-#base = 16
-#length = 2
-#choice = 5
-#epochs = 20
-#dimension = 100
-#training_size = 5000
-#testing_size = 100
-#alphabet = ALPHABET[:base]
-
-#g = Generator(alphabet, length, choice)
-#train_data = g.generate_data(training_size)
-#test_data = g.generate_data(testing_size)
-#trainer = Trainer(train_data, test_data, epochs, dimension)
-#model = trainer.train()
-#model = trainer.batch_train(1000)
-#train_acc = trainer.evaluate(model, train_data[:200])
-#test_acc = trainer.evaluate(model, test_data)
-
-#base = 16
-#length = 2
-#choice = 5
-#epochs = 2000
-#dimension = 100
-#training_size = 2500
-#testing_size = 100
-#alphabet = ALPHABET[:base]
-
-#g = Generator(alphabet, length, choice)
-#train_data = g.generate_data(training_size)
-#test_data = g.generate_data(testing_size)
-#trainer = Trainer(train_data, test_data, epochs, dimension)
-#old_model = MultiLayerClassifier.initialize_from_model_and_vocab(model, trainer.vocab)
-#new_model = trainer.batch_train(old_model)
-#train_acc = trainer.evaluate(new_model, train_data[:1000])
-#test_acc = trainer.evaluate(new_model, test_data)
-#print()
-#print(train_acc)
-#print(test_acc)
